# api-contract/openapi-admin.json
{"components":{"schemas":{"AdjustRequest":{"properties":{"delta":{"format":"int32","type":"integer"},"reason":{"maxLength":255,"minLength":0,"type":"string"}},"required":["delta","reason"],"type":"object"},"AdminOrderDetail":{"properties":{"createdAt":{"format":"date-time","type":"string"},"currency":{"type":"string"},"items":{"items":{"$ref":"#/components/schemas/OrderItemResponse"},"type":"array"},"orderId":{"type":"string"},"shipments":{"items":{"$ref":"#/components/schemas/ShipmentInfo"},"type":"array"},"shippingAddress":{"$ref":"#/components/schemas/ShippingAddressInfo"},"status":{"description":"Order status lifecycle","enum":["PENDING_PAYMENT","PAID","SHIPPED","PAYMENT_FAILED","REFUND_PENDING","REFUNDED","PLACED","CANCELLED","FULFILLED"],"type":"string"},"totalAmount":{"type":"number"},"userEmail":{"type":"string"}},"type":"object"},"ApiProblem":{"properties":{"code":{"example":"VALIDATION_FAILED","type":"string"},"detail":{"example":"Validation failed","type":"string"},"instance":{"example":"/api/v1/products","type":"string"},"requestId":{"example":"3f9d...","type":"string"},"retryAfterSeconds":{"example":12,"format":"int32","type":"integer"},"status":{"example":400,"format":"int32","type":"integer"},"timestamp":{"example":"2026-02-17T12:34:56Z","type":"string"},"title":{"example":"Validation error","type":"string"},"type":{"example":"urn:problem:validation","type":"string"},"violations":{"items":{"properties":{"field":{"type":"string"},"invalidValue":{"type":"object"},"message":{"type":"string"},"path":{"type":"string"},"rejectedValue":{"type":"object"}},"type":"object"},"type":"array"}},"type":"object"},"CancelRequest":{"properties":{"reason":{"maxLength":255,"minLength":0,"type":"string"}},"type":"object"},"OrderItemResponse":{"properties":{"currency":{"type":"string"},"lineTotal":{"type":"number"},"name":{"type":"string"},"productId":{"format":"int64","type":"integer"},"quantity":{"format":"int32","type":"integer"},"sku":{"type":"string"},"unitPrice":{"type":"number"}},"type":"object"},"PageMeta":{"properties":{"first":{"example":true,"type":"boolean"},"last":{"example":false,"type":"boolean"},"number":{"example":0,"format":"int32","type":"integer"},"size":{"example":20,"format":"int32","type":"integer"},"totalElements":{"example":120,"format":"int64","type":"integer"},"totalPages":{"example":6,"format":"int32","type":"integer"}},"type":"object"},"PageResponse":{"properties":{"items":{"description":"Listed items","items":{"$ref":"#/components/schemas/WebhookEventRow"},"type":"array"},"page":{"$ref":"#/components/schemas/PageMeta","description":"Paging metadata"}},"type":"object"},"RestockRequest":{"properties":{"quantity":{"format":"int32","minimum":1,"type":"integer"},"reason":{"maxLength":255,"minLength":0,"type":"string"}},"required":["reason"],"type":"object"},"ShipRequest":{"properties":{"carrier":{"maxLength":64,"minLength":0,"type":"string"},"trackingNumber":{"maxLength":128,"minLength":0,"type":"string"}},"type":"object"},"ShipmentInfo":{"properties":{"carrier":{"type":"string"},"deliveredAt":{"format":"date-time","type":"string"},"shipmentId":{"type":"string"},"shippedAt":{"format":"date-time","type":"string"},"status":{"type":"string"},"trackingNumber":{"type":"string"}},"type":"object"},"ShipmentResponse":{"properties":{"carrier":{"type":"string"},"deliveredAt":{"format":"date-time","type":"string"},"shipmentId":{"type":"string"},"shippedAt":{"format":"date-time","type":"string"},"status":{"type":"string"},"trackingNumber":{"type":"string"}},"type":"object"},"ShippingAddressInfo":{"properties":{"city":{"type":"string"},"countryCode":{"type":"string"},"fullName":{"type":"string"},"line1":{"type":"string"},"line2":{"type":"string"},"phone":{"type":"string"},"postalCode":{"type":"string"},"state":{"type":"string"}},"type":"object"},"StockResponse":{"properties":{"stockQuantity":{"format":"int32","type":"integer"}},"type":"object"},"WebhookEventRow":{"properties":{"eventId":{"type":"string"},"payloadHash":{"type":"string"},"provider":{"type":"string"},"receivedAt":{"format":"date-time","type":"string"}},"type":"object"}},"securitySchemes":{"bearerAuth":{"bearerFormat":"JWT","scheme":"bearer","type":"http"}}},"info":{"description":"E-commerce backend API (Spring Boot 4)","title":"Pehlione Web API","version":"v1"},"openapi":"3.1.0","paths":{"/api/v1/admin/inventory/products/{productId}/adjust":{"post":{"operationId":"adjust","parameters":[{"in":"path","name":"productId","required":true,"schema":{"format":"int64","type":"integer"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdjustRequest"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/StockResponse"}}},"description":"OK"},"400":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Bad Request / Validation"},"401":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Unauthorized"},"403":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Forbidden"},"409":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Conflict"},"429":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Too Many Requests"},"500":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Internal Server Error"}},"security":[{"bearerAuth":[]}],"tags":["Admin - Inventory"]}},"/api/v1/admin/inventory/products/{productId}/restock":{"post":{"operationId":"restock","parameters":[{"in":"path","name":"productId","required":true,"schema":{"format":"int64","type":"integer"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestockRequest"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/StockResponse"}}},"description":"OK"},"400":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Bad Request / Validation"},"401":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Unauthorized"},"403":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Forbidden"},"409":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Conflict"},"429":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Too Many Requests"},"500":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Internal Server Error"}},"security":[{"bearerAuth":[]}],"tags":["Admin - Inventory"]}},"/api/v1/admin/orders":{"get":{"description":"Filter by status/email/orderId/date range.","operationId":"list_2","parameters":[{"description":"Order status filter","in":"query","name":"status","required":false,"schema":{"description":"Order status lifecycle","enum":["PENDING_PAYMENT","PAID","SHIPPED","PAYMENT_FAILED","REFUND_PENDING","REFUNDED","PLACED","CANCELLED","FULFILLED"],"type":"string"}},{"description":"User email contains","example":"gmail.com","in":"query","name":"email","required":false,"schema":{"type":"string"}},{"description":"Order id contains","example":"6b7a","in":"query","name":"q","required":false,"schema":{"type":"string"}},{"description":"Created-at lower bound (ISO date-time)","example":"2026-02-01T00:00:00Z","in":"query","name":"from","required":false,"schema":{"format":"date-time","type":"string"}},{"description":"Created-at upper bound (ISO date-time)","example":"2026-02-28T23:59:59Z","in":"query","name":"to","required":false,"schema":{"format":"date-time","type":"string"}},{"description":"Zero-based page index (0..N)","in":"query","name":"page","required":false,"schema":{"default":0,"minimum":0,"type":"integer"}},{"description":"The size of the page to be returned","in":"query","name":"size","required":false,"schema":{"default":20,"minimum":1,"type":"integer"}},{"description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","in":"query","name":"sort","required":false,"schema":{"items":{"type":"string"},"type":"array"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponse"}}},"description":"OK"},"400":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Bad Request / Validation"},"401":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Unauthorized"},"403":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Forbidden"},"409":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Conflict"},"429":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Too Many Requests"},"500":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Internal Server Error"}},"security":[{"bearerAuth":[]}],"summary":"Admin: list orders","tags":["Admin - Orders"]}},"/api/v1/admin/orders/{orderId}":{"get":{"operationId":"get","parameters":[{"in":"path","name":"orderId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/AdminOrderDetail"}}},"description":"OK"},"400":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Bad Request / Validation"},"401":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Unauthorized"},"403":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Forbidden"},"409":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Conflict"},"429":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Too Many Requests"},"500":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Internal Server Error"}},"security":[{"bearerAuth":[]}],"tags":["Admin - Orders"]}},"/api/v1/admin/orders/{orderId}/cancel":{"post":{"operationId":"cancel","parameters":[{"in":"path","name":"orderId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CancelRequest"}}}},"responses":{"200":{"description":"OK"},"400":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Bad Request / Validation"},"401":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Unauthorized"},"403":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Forbidden"},"409":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Conflict"},"429":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Too Many Requests"},"500":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Internal Server Error"}},"security":[{"bearerAuth":[]}],"tags":["Admin - Fulfillment"]}},"/api/v1/admin/orders/{orderId}/deliver":{"post":{"operationId":"deliver","parameters":[{"in":"path","name":"orderId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"},"400":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Bad Request / Validation"},"401":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Unauthorized"},"403":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Forbidden"},"409":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Conflict"},"429":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Too Many Requests"},"500":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Internal Server Error"}},"security":[{"bearerAuth":[]}],"tags":["Admin - Fulfillment"]}},"/api/v1/admin/orders/{orderId}/ship":{"post":{"operationId":"ship","parameters":[{"in":"path","name":"orderId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShipRequest"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ShipmentResponse"}}},"description":"OK"},"400":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Bad Request / Validation"},"401":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Unauthorized"},"403":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Forbidden"},"409":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Conflict"},"429":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Too Many Requests"},"500":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Internal Server Error"}},"security":[{"bearerAuth":[]}],"tags":["Admin - Fulfillment"]}},"/api/v1/admin/refunds":{"get":{"description":"Filter by status, user email and order id.","operationId":"list_1","parameters":[{"description":"Refund status filter","in":"query","name":"status","required":false,"schema":{"description":"Refund processing status","enum":["PENDING","SUCCEEDED","FAILED","CANCELLED"],"type":"string"}},{"description":"User email contains","example":"gmail.com","in":"query","name":"email","required":false,"schema":{"type":"string"}},{"description":"Order id exact match","example":"6b7a2f42","in":"query","name":"orderId","required":false,"schema":{"type":"string"}},{"description":"Zero-based page index (0..N)","in":"query","name":"page","required":false,"schema":{"default":0,"minimum":0,"type":"integer"}},{"description":"The size of the page to be returned","in":"query","name":"size","required":false,"schema":{"default":20,"minimum":1,"type":"integer"}},{"description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","in":"query","name":"sort","required":false,"schema":{"items":{"type":"string"},"type":"array"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponse"}}},"description":"OK"},"400":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Bad Request / Validation"},"401":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Unauthorized"},"403":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Forbidden"},"409":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Conflict"},"429":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Too Many Requests"},"500":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Internal Server Error"}},"security":[{"bearerAuth":[]}],"summary":"Admin: list refunds","tags":["Admin - Refunds"]}},"/api/v1/admin/webhook-events":{"get":{"description":"Filter events by provider and event id.","operationId":"list","parameters":[{"description":"Payment provider","example":"mock","in":"query","name":"provider","required":false,"schema":{"type":"string"}},{"description":"Event id contains","example":"evt_","in":"query","name":"eventId","required":false,"schema":{"type":"string"}},{"description":"Zero-based page index (0..N)","in":"query","name":"page","required":false,"schema":{"default":0,"minimum":0,"type":"integer"}},{"description":"The size of the page to be returned","in":"query","name":"size","required":false,"schema":{"default":20,"minimum":1,"type":"integer"}},{"description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","in":"query","name":"sort","required":false,"schema":{"items":{"type":"string"},"type":"array"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponse"}}},"description":"OK"},"400":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Bad Request / Validation"},"401":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Unauthorized"},"403":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Forbidden"},"409":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Conflict"},"429":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Too Many Requests"},"500":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ApiProblem"}}},"description":"Internal Server Error"}},"security":[{"bearerAuth":[]}],"summary":"Admin: list webhook events","tags":["Admin - Webhook Events"]}}},"security":[{"bearerAuth":[]}],"tags":[{"description":"Administrative stock management endpoints","name":"Admin - Inventory"},{"description":"Administrative refund listing endpoints","name":"Admin - Refunds"},{"description":"Administrative webhook event query endpoints","name":"Admin - Webhook Events"},{"description":"Administrative order management endpoints","name":"Admin - Orders"},{"description":"Administrative shipment and fulfillment endpoints","name":"Admin - Fulfillment"}]}